requirements.txt
google-ai-generativelanguage>=0.5.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.28.0 